+++ protocols/request-response/src/lib.rs
@@ -58,31 +58,15 @@
 
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-mod codec_priv;
-#[deprecated(
-    note = "The `codec` module will be made private in the future and should not be depended on."
-)]
-pub mod codec {
-    pub use super::codec_priv::*;
-}
+mod codec;
+mod handler;
 
-mod handler_priv;
-#[deprecated(
-    note = "The `handler` module will be made private in the future and should not be depended on."
-)]
-pub mod handler {
-    pub use super::handler_priv::*;
-}
+pub use codec::{Codec, ProtocolName};
+pub use handler::ProtocolSupport;
 
-pub use codec_priv::{Codec, ProtocolName};
-
-#[allow(deprecated)]
-pub use codec_priv::RequestResponseCodec;
-
-pub use handler_priv::ProtocolSupport;
-
+use crate::handler::protocol::RequestProtocol;
 use futures::channel::oneshot;
-use handler_priv::{Handler, RequestProtocol};
+use handler::Handler;
 use libp2p_core::{ConnectedPoint, Endpoint, Multiaddr};
 use libp2p_identity::PeerId;
 use libp2p_swarm::{
@@ -99,37 +83,6 @@ use std::{
     task::{Context, Poll},
     time::Duration,
 };
-
-#[deprecated(
-    since = "0.24.0",
-    note = "Use libp2p::request_response::Behaviour instead."
-)]
-pub type RequestResponse<TCodec> = Behaviour<TCodec>;
-
-#[deprecated(
-    since = "0.24.0",
-    note = "Use re-exports that omit `RequestResponse` prefix, i.e. `libp2p::request_response::Config`"
-)]
-pub type RequestResponseConfig = Config;
-
-#[deprecated(
-    since = "0.24.0",
-    note = "Use re-exports that omit `RequestResponse` prefix, i.e. `libp2p::request_response::Event`"
-)]
-pub type RequestResponseEvent<TRequest, TResponse> = Event<TRequest, TResponse>;
-
-#[deprecated(
-    since = "0.24.0",
-    note = "Use re-exports that omit `RequestResponse` prefix, i.e. `libp2p::request_response::Message`"
-)]
-pub type RequestResponseMessage<TRequest, TResponse, TChannelResponse> =
-    Message<TRequest, TResponse, TChannelResponse>;
-
-#[deprecated(
-    since = "0.24.0",
-    note = "Use re-exports that omit `RequestResponse` prefix, i.e. `libp2p::request_response::handler::Event`"
-)]
-pub type HandlerEvent<TCodec> = handler_priv::Event<TCodec>;
 
 /// An inbound request or response.
 #[derive(Debug)]
@@ -815,7 +768,7 @@ where
         event: THandlerOutEvent<Self>,
     ) {
         match event {
-            handler_priv::Event::Response {
+            handler::Event::Response {
                 request_id,
                 response,
             } => {
@@ -832,7 +785,7 @@ where
                 self.pending_events
                     .push_back(ToSwarm::GenerateEvent(Event::Message { peer, message }));
             }
-            handler_priv::Event::Request {
+            handler::Event::Request {
                 request_id,
                 request,
                 sender,
@@ -863,7 +816,7 @@ where
                     }
                 }
             }
-            handler_priv::Event::ResponseSent(request_id) => {
+            handler::Event::ResponseSent(request_id) => {
                 let removed = self.remove_pending_outbound_response(&peer, connection, request_id);
                 debug_assert!(
                     removed,
@@ -876,7 +829,7 @@ where
                         request_id,
                     }));
             }
-            handler_priv::Event::ResponseOmission(request_id) => {
+            handler::Event::ResponseOmission(request_id) => {
                 let removed = self.remove_pending_outbound_response(&peer, connection, request_id);
                 debug_assert!(
                     removed,
@@ -890,7 +843,7 @@ where
                         error: InboundFailure::ResponseOmission,
                     }));
             }
-            handler_priv::Event::OutboundTimeout(request_id) => {
+            handler::Event::OutboundTimeout(request_id) => {
                 let removed = self.remove_pending_inbound_response(&peer, connection, &request_id);
                 debug_assert!(
                     removed,
@@ -904,7 +857,7 @@ where
                         error: OutboundFailure::Timeout,
                     }));
             }
-            handler_priv::Event::InboundTimeout(request_id) => {
+            handler::Event::InboundTimeout(request_id) => {
                 // Note: `Event::InboundTimeout` is emitted both for timing
                 // out to receive the request and for timing out sending the response. In the former
                 // case the request is never added to `pending_outbound_responses` and thus one can
@@ -918,7 +871,7 @@ where
                         error: InboundFailure::Timeout,
                     }));
             }
-            handler_priv::Event::OutboundUnsupportedProtocols(request_id) => {
+            handler::Event::OutboundUnsupportedProtocols(request_id) => {
                 let removed = self.remove_pending_inbound_response(&peer, connection, &request_id);
                 debug_assert!(
                     removed,
@@ -932,7 +885,7 @@ where
                         error: OutboundFailure::UnsupportedProtocols,
                     }));
             }
-            handler_priv::Event::InboundUnsupportedProtocols(request_id) => {
+            handler::Event::InboundUnsupportedProtocols(request_id) => {
                 // Note: No need to call `self.remove_pending_outbound_response`,
                 // `Event::Request` was never emitted for this request and
                 // thus request was never added to `pending_outbound_responses`.