+++ protocols/request-response/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Raise MSRV to 1.65.
   See [PR 3715].
+- Remove deprecated `RequestResponse` prefixed items. See [PR 3702].
 
 - Remove `InboundFailure::UnsupportedProtocols`.
   This variant is no longer constructed.
@@ -9,6 +10,7 @@
 
 [PR 3605]: https://github.com/libp2p/rust-libp2p/pull/3605
 [PR 3715]: https://github.com/libp2p/rust-libp2p/pull/3715
+[PR 3702]: https://github.com/libp2p/rust-libp2p/pull/3702
 
 ## 0.24.1